+++ samplesheets/templatetags/samplesheets_tags.py
@@ -1,11 +1,20 @@
+import random
+import string
+
 from django import template
 from django.urls import reverse
 
 from ..models import Investigation, Study, Assay, GenericMaterial, \
     GENERIC_MATERIAL_TYPES
-from ..rendering import SampleSheetHTMLRenderer as Renderer
 
 register = template.Library()
+
+
+# Local constants
+EMPTY_VALUE = '-'
+
+
+# General tags -----------------------------------------------------------------
 
 
 @register.simple_tag
@@ -16,42 +25,6 @@ def get_investigation(project):
 
     except Investigation.DoesNotExist:
         return None
-
-
-@register.simple_tag
-def render_cell(cell):
-    """Return assay table cell as HTML"""
-    return Renderer.render_cell(cell)
-
-
-@register.simple_tag
-def render_links_cell(row):
-    """Render iRODS/IGV links cell"""
-    return Renderer.render_links_cell()
-
-
-@register.simple_tag
-def render_top_header(section):
-    """Render section of top header"""
-    return Renderer.render_top_header(section)
-
-
-@register.simple_tag
-def render_links_top_header():
-    """Render top links header"""
-    return Renderer.render_links_top_header()
-
-
-@register.simple_tag
-def render_header(header):
-    """Render section of top header"""
-    return Renderer.render_header(header)
-
-
-@register.simple_tag
-def render_links_header():
-    """Render links column header"""
-    return Renderer.render_links_header()
 
 
 @register.simple_tag
@@ -162,3 +135,113 @@ def get_assay_info_html(assay):
 
     ret += '</div>\n'
     return ret
+
+
+# Rendering tags ---------------------------------------------------------------
+
+
+@register.simple_tag
+def render_top_header(section):
+    """
+    Render section of top header
+    :param section: Header section (dict)
+    :return: String (contains HTML)
+    """
+    return '<th class="bg-{} text-nowrap text-white omics-ss-top-header" ' \
+           'colspan="{}" original-colspan="{}" {}>{}</th>\n'.format(
+            section['colour'],
+            section['colspan'],     # Actual colspan
+            section['colspan'],     # Original colspan
+            ''.join(['{}-cols="{}" '.format(k, v) for
+                     k, v in section['hiding'].items()]),
+            section['value'])
+
+
+@register.simple_tag
+def get_random_id():
+    """
+    Return random string for link ids
+    :return: string
+    """
+    return ''.join(random.SystemRandom().choice(
+        string.ascii_lowercase + string.digits) for x in range(16))
+
+
+@register.simple_tag
+def render_cell(cell):
+    """
+    Return data table cell as HTML
+    :param cell: Cell dict
+    :return: String (contains HTML)
+    """
+    td_class_str = ' '.join(cell['classes'])
+
+    # If repeating cell, return that
+    if cell['repeat']:
+        return '<td class="bg-light text-muted text-center {}">' \
+               '"</td>\n'.format(td_class_str)
+
+    # Right aligning
+    def is_num(x):
+        try:
+            float(x)
+            return True
+
+        except ValueError:
+            return False
+
+    if cell['value'] and is_num(cell['value']):
+        td_class_str += ' text-right'
+
+    # Build <td>
+    ret = '<td '
+
+    # Add extra attrs if present
+    if cell['attrs']:
+        for k, v in cell['attrs'].items():
+            ret += '{}="{}" '.format(k, v)
+
+    if cell['tooltip']:
+        ret += 'class="{}" title="{}" data-toggle="tooltip" ' \
+               'data-placement="top">'.format(td_class_str, cell['tooltip'])
+
+    else:
+        ret += 'class="{}">'.format(td_class_str)
+
+    if cell['value']:
+        if cell['link']:
+            ret += '<a href="{}" target="_blank">{}</a>'.format(
+                cell['link'], cell['value'])
+
+        else:
+            ret += cell['value']
+
+        if cell['unit']:
+            ret += '&nbsp;<span class=" text-muted">{}</span>'.format(
+                cell['unit'])
+
+    else:   # Empty value
+        ret += EMPTY_VALUE
+
+    ret += '</td>\n'
+    return ret
+
+@register.simple_tag
+def render_links_cell(row):
+    """
+    Return links cell for row as HTML
+    :return: String (contains HTML)
+    """
+    # TODO: Add actual links
+    # TODO: Refactor/cleanup, this is a quick screenshot HACK
+
+    return '<td class="bg-light omics-ss-data-cell-links">\n' \
+           '  <div class="btn-group omics-ss-data-btn-group">\n' \
+           '    <button class="btn btn-secondary dropdown-toggle btn-sm ' \
+           '                   omics-ss-data-dropdown"' \
+           '                   type="button" data-toggle="dropdown" ' \
+           '                   aria-expanded="false">' \
+           '                   <i class="fa fa-external-link"></i>' \
+           '    </button>' \
+           '  </div>\n' \
+           '</td>\n'